chore: refactor namespaces and update converter references to DataTemplates.
