+++ demo/Ursa.Demo/Pages/BreadcrumbDemo.axaml
@@ -6,17 +6,13 @@
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:u="https://irihi.tech/ursa"
     xmlns:vm="using:Ursa.Demo.ViewModels"
-    xmlns:converters="clr-namespace:Ursa.Demo.Converters"
+    xmlns:converters="clr-namespace:Ursa.Demo.DataTemplates"
     d:DesignHeight="450"
     d:DesignWidth="800"
     x:CompileBindings="True"
     x:DataType="vm:BreadcrumbDemoViewModel"
     mc:Ignorable="d">
-    <UserControl.Resources>
-        <converters:IconNameToPathConverter x:Key="IconConverter" />
-    </UserControl.Resources>
     <StackPanel>
-        
         <u:Breadcrumb>
             <TextBlock Text="Hello" />
             <u:BreadcrumbItem Content="World" IsReadOnly="True" />
@@ -50,7 +46,7 @@
                 <DataTemplate DataType="x:String">
                     <PathIcon
                         Theme="{StaticResource InnerPathIcon}"
-                        Data="{Binding Converter={StaticResource IconConverter}}" />
+                        Data="{Binding Converter={converters:IconNameToPathConverter}}" />
                 </DataTemplate>
             </u:Breadcrumb.IconTemplate>
         </u:Breadcrumb>
