+++ demo/Ursa.Demo/Pages/FormDemo.axaml
@@ -6,7 +6,7 @@
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:u="https://irihi.tech/ursa"
     xmlns:vm="clr-namespace:Ursa.Demo.ViewModels;assembly=Ursa.Demo"
-    xmlns:converters="clr-namespace:Ursa.Demo.Converters"
+    xmlns:dataTemplates="clr-namespace:Ursa.Demo.DataTemplates"
     d:DesignHeight="450"
     d:DesignWidth="800"
     x:CompileBindings="True"
@@ -62,7 +62,7 @@
                     </Style>
                 </u:Form.Styles>
                 <u:Form.ItemTemplate>
-                    <converters:FormDataTemplateSelector>
+                    <dataTemplates:FormDataTemplateSelector>
                         <DataTemplate x:Key="{x:Type vm:FormTextViewModel}" DataType="vm:FormTextViewModel">
                             <TextBox Text="{Binding Value}"/>
                         </DataTemplate>
@@ -72,7 +72,7 @@
                         <DataTemplate x:Key="{x:Type vm:FormDateRangeViewModel}" DataType="vm:FormDateRangeViewModel">
                             <u:DateRangePicker SelectedStartDate="{Binding Start}" SelectedEndDate="{Binding End}"/>
                         </DataTemplate>
-                    </converters:FormDataTemplateSelector>
+                    </dataTemplates:FormDataTemplateSelector>
                 </u:Form.ItemTemplate>
             </u:Form>
         </StackPanel>
